Merge typed defaults if prop not present
Typed buckets do not automatically inherit newly added default
properties (old-skool default buckets do thanks to the existence of
bucket fixups.)

As per the bug report, when PUTing to a Typed bucket after upgrade to
2.2.5 the call to riak_kv_util:expand_rw_value/4 throws if the
property `node_confirms` is not in the bucket properties, and for a
typed bucket created and activated before upgrade it won't be.

This is somewhat contentious. Ex-Basho @seancribbs has pointed out
that there is a reason for that throw, and that all nodes should agree
on the value, and misconfiguration can be an issue here.  There are
perhpas more complexities (we're retroactivley adding a reserved word,
for example.) Caveats in docs and probability should guard against
that latter though.

Given that the reasons not to do this are lost to time, and the
alternatives unpleasant, I'm content that this is a reasonable way for
typed buckets to inherit **NEW** properties added as defaults after
typed bucket creation.

There is an associated riak_core commit.